Add ability to relocate CDR forcing to wet points.
Sometimes parent and child grids don't agree on the locations of land and
water points, so the CDR forcing module would occasionally find that
the nearest grid point is dry and would error out.  This PR adds the option
to relocate the CDR to the nearest wet point instead, so that it doesn't
error out.